No lp_release_parport while write is going on
This patch was done by IBM a while back, but apparently never made it
into mainline. It fixes a problem in the lp driver that can cause oopses.

Scenario:
 process A:	calls lp_write, which in turn calls parport_ieee1284_write_compat,
 		and that invokes parport_wait_peripheral
 process B:	meanwhile does an ioctl(LPGETSTATUS), which call lp_release_parport
 		when done. This function will set physport->cad = NULL.
 process A:	parport_wait_peripheral tries to dereference physport->cad and
 		dies

The patch below simply protects that code with the port_mutex in order to
protect against simultaneous calls to lp_read/lp_write.

Similar protection is probably required for ioctl(LPRESET).

Signed-off-by: okir@suse.de
Signed-off-by: redstar3894 <redstar3894@gmail.com>