ia64: Call migration code on correctable errors v8
Migrate data off pages with correctable memory errors.  This patch is the
ia64 specific piece.  It connects the CPE handler to the page migration
code.  It is implemented as a kernel loadable module, similar to the mca
recovery code (mca_recovery.ko).  This allows the feature to be turned off
by uninstalling the module.

Update Jan 19 2009 jeffm:
- isolate_lru_page doesn't put the page on a list anymore

Signed-off-by: Russ Anderson <rja@sgi.com>
Signed-off-by: redstar3894 <redstar3894@gmail.com>