add syslog printing to xmon debugger.
This patch 'dmesg'/printk log buffer printing to xmon.  I find this
useful because crashes are almost always preceeded by interesting
printk's.   This patch is simple & straightforward, except for one
possibly controversial aspect: it embeds a small snippet in
kernel/printk.c to return the location of the syslog.  This is
needed because kallsyms and even CONFIG_KALLSYMS_ALL is not enough
to reveal the location of log_buf.   This code is about 90%
cut-n-paste of earlier code from Keith Owens.

Signed-off-by: Olaf Hering <olh@suse.de>
Signed-off-by: redstar3894 <redstar3894@gmail.com>